refactor: moved files to align with ether changes
